requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
google-generativeai==0.7.2
supabase==2.6.0
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.2